requirements.txt
python-dotenv
langchain
openai
langchain-openai
langchain-community==0.2.15
openai
requests
telethon
tiktoken
lxml==5.1.0
matplotlib
pandas==1.3.5
sqlalchemy~=1.4.50
flask